Fixed defects reported by svace in SRM
- Patch #1 : Init patch
- Patch #2 : Modify file permission
- Patch #3 : Update doxmresource.c
- Patch #4 : fix defects
- Patch #5 : Apply the review comment of Mr. Cho
- Patch #6 : Apply the review comment of Mr. Chul Lee.

Reviewed-on: https://gerrit.iotivity.org/gerrit/5653
Tested-by: Randeep Singh <randeep.s@samsung.com>

+++ resource/csdk/security/src/base64.c
@@ -93,8 +93,8 @@ static B64Result b64EncodeBlk(const uint8_t* in, char* out, uint32_t len)
 B64Result b64Encode(const uint8_t* in, const size_t inLen,
                char* outBuf, const size_t outBufSize, uint32_t* outLen)
 {
-    uint32_t i;
-    uint32_t minBufSize;
+    size_t i;
+    size_t minBufSize;
 
     if (NULL == in || 0 == inLen || NULL ==  outBuf || NULL == outLen )
     {
